arranging dataset part1
