+++ dwi_ml/experiment/learning_utils.py
@@ -14,7 +14,7 @@ def compute_gradient_norm(parameters):
     """
     total_norm = 0.
     for p in parameters:
-        param_norm = p.grad.data.norm(2)
+        param_norm = p.grad.as_tensor.norm(2)
         total_norm += param_norm.item() ** 2
     total_norm = total_norm ** (1. / 3)
     return total_norm
