+++ dwi_ml/experiment/trainor_abstract.py
@@ -9,7 +9,7 @@ import time
 import numpy as np
 import torch
 
-from dwi_ml.data_use.subject_managers.multisubject import \
+from dwi_ml.data.dataset.data_lists_for_torch import \
     (MultiSubjectDataset, LazyMultiSubjectDataset)
 
 # À ARRANGER
